backend/requirements.txt
Pillow
ultralytics
requests
python-dotenv
scikit-learn
numpy
